+++ lib/src/wui_builder/create_nodes.dart
@@ -1,38 +1,17 @@
-part of wui_builder;
+import 'dart:html';
+
+import 'component.dart';
+import 'velement.dart';
+import 'vnode.dart';
 
 void render(VNode vnode, Element mount) {
-  mount.append(_createNode(vnode));
+  mount.append(createNode(vnode));
 }
 
-Element _createNode(VNode vnode) {
+Element createNode(VNode vnode) {
   if (vnode.vNodeType == VNodeTypes.Element) {
-    return _createElementNode(vnode as VElement);
+    return createElementNode(vnode as VElement);
   } else {
-    return _createComponentNode(vnode as Component);
+    return createComponentNode(vnode as Component);
   }
-}
-
-Element _createElementNode(VElement vnode) {
-  final Element domNode = vnode.elementFactory();
-  vnode.ref = domNode;
-  vnode.applyAttributesToElement(domNode);
-  if (vnode.shouldUpdateSubs) vnode.applyEventListenersToElement(domNode);
-  if (vnode.children != null) {
-    for (final c in vnode.children) {
-      domNode.append(_createNode(c));
-      c.parent = vnode;
-    }
-  }
-  return domNode;
-}
-
-Element _createComponentNode(Component vnode) {
-  vnode._state = vnode.getInitialState();
-  vnode.componentWillMount();
-  vnode._render();
-  vnode._renderResult.parent = vnode;
-  final domNode = _createNode(vnode._renderResult);
-  vnode.ref = domNode;
-  vnode.componentDidMount();
-  return domNode;
 }
