+++ lib/src/wui_builder/update_tracker.dart
@@ -1,42 +1,44 @@
-part of wui_builder;
+import 'dart:html';
 
-// a _UpdateTracker contains state relating to where in the
+import 'component.dart';
+import 'cursors.dart';
+import 'vnode.dart';
+import 'update_queue.dart';
+
+// a UpdateTracker contains state relating to where in the
 // reconciliation process we are for a given update. It tracks
 // what work still needs to be done for a given update
 // in a way that allows reconciliation to be paused and resumed.
-class _UpdateTracker {
+class UpdateTracker {
   // current location in the update
-  _Cursor cursor;
+  Cursor cursor;
   final bool isAsync;
-  bool hasStarted;
   bool isCancelled = false;
-  List<_PendingCursor> pendingCursors = new List<_PendingCursor>();
+  List<PendingCursor> pendingCursors = new List<PendingCursor>();
   IdleDeadline deadline;
 
   // used to reset state if update is cancelled
   final dynamic prevState;
   final Component updatingComponent;
 
-  _UpdateTracker.sync(Element node, Component newVNode, this.prevState)
-      : cursor = new _Cursor(
+  UpdateTracker.sync(Element node, Component newVNode, this.prevState)
+      : cursor = new Cursor(
           node.parent,
           node,
           newVNode,
           newVNode,
         ),
         isAsync = false,
-        hasStarted = true,
         updatingComponent = newVNode;
 
-  _UpdateTracker.async(Element node, Component newVNode, this.prevState)
-      : cursor = new _Cursor(
+  UpdateTracker.async(Element node, Component newVNode, this.prevState)
+      : cursor = new Cursor(
           node.parent,
           node,
           newVNode,
           newVNode,
         ),
         isAsync = true,
-        hasStarted = false,
         updatingComponent = newVNode;
 
   // update changes the current location of the update
@@ -54,23 +56,22 @@ class _UpdateTracker {
     if (!isAsync) return false;
     if (_isPaused) return _isPaused;
     _isPaused = deadline.timeRemaining() < 1;
-    if (_isPaused) _queueProcessingUpdate(this);
+    if (_isPaused) queueProcessingUpdate(this);
     return _isPaused;
   }
 
   void cancel() {
     // revert the state change.
-    updatingComponent._state = prevState;
+    revertState(updatingComponent, prevState);
     isCancelled = true;
   }
 
   void refresh(IdleDeadline d) {
-    hasStarted = true;
     deadline = d;
     _isPaused = false;
   }
 
-  void pushPendingCursor(_PendingCursor cursor) {
+  void pushPendingCursor(PendingCursor cursor) {
     pendingCursors.add(cursor);
   }
