+++ lib/src/wui_builder/velement.dart
@@ -1,16 +1,17 @@
-library component;
+library velement;
 
 import 'dart:html';
+import 'dart:async';
 
 import 'package:meta/meta.dart';
 
 import 'create_nodes.dart';
 import 'cursors.dart';
 import 'update_processor.dart';
-import 'update_queue.dart';
 import 'update_tracker.dart';
 import 'vnode.dart';
 
-part 'component/component.dart';
-part 'component/create.dart';
-part 'component/update.dart';
+part 'velement/create.dart';
+part 'velement/typdefs.dart';
+part 'velement/update.dart';
+part 'velement/velement.dart';
