+++ lib/src/wui_builder/velement/typdefs.dart
@@ -1,0 +1,5 @@
+part of velement;
+
+typedef void EventHandler<T>(T event);
+
+typedef void StyleBuilder(CssStyleDeclaration style);
