+++ lib/vsvg.dart
@@ -1,16 +1,19 @@
 import 'dart:svg';
-import 'vhtml.dart' show VElement;
+import 'package:meta/meta.dart';
+import 'wui_builder.dart' show VElement;
 
 class VAElement extends VGraphicsElement<AElement> {
   @override
   AElement elementFactory() => new AElement();
 
   @override
+  @protected
   void applyAttributesToElement(AElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VAElement prev, AElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -21,11 +24,13 @@ class VAnimateElement extends VAnimationElement<AnimateElement> {
   AnimateElement elementFactory() => new AnimateElement();
 
   @override
+  @protected
   void applyAttributesToElement(AnimateElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VAnimateElement prev, AnimateElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -36,11 +41,13 @@ class VAnimateMotionElement extends VAnimationElement<AnimateMotionElement> {
   AnimateMotionElement elementFactory() => new AnimateMotionElement();
 
   @override
+  @protected
   void applyAttributesToElement(AnimateMotionElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VAnimateMotionElement prev, AnimateMotionElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -53,11 +60,13 @@ class VAnimateTransformElement
   AnimateTransformElement elementFactory() => new AnimateTransformElement();
 
   @override
+  @protected
   void applyAttributesToElement(AnimateTransformElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VAnimateTransformElement prev, AnimateTransformElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -67,11 +76,13 @@ class VAnimateTransformElement
 abstract class VAnimationElement<T extends AnimationElement>
     extends VSvgElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VAnimationElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -82,11 +93,13 @@ class VCircleElement extends VGeometryElement<CircleElement> {
   CircleElement elementFactory() => new CircleElement();
 
   @override
+  @protected
   void applyAttributesToElement(CircleElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VCircleElement prev, CircleElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -97,11 +110,13 @@ class VClipPathElement extends VGraphicsElement<ClipPathElement> {
   ClipPathElement elementFactory() => new ClipPathElement();
 
   @override
+  @protected
   void applyAttributesToElement(ClipPathElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VClipPathElement prev, ClipPathElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -112,11 +127,13 @@ class VDefsElement extends VGraphicsElement<DefsElement> {
   DefsElement elementFactory() => new DefsElement();
 
   @override
+  @protected
   void applyAttributesToElement(DefsElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VDefsElement prev, DefsElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -127,11 +144,13 @@ class VDescElement extends VSvgElement<DescElement> {
   DescElement elementFactory() => new DescElement();
 
   @override
+  @protected
   void applyAttributesToElement(DescElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VDescElement prev, DescElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -140,11 +159,13 @@ class VDescElement extends VSvgElement<DescElement> {
 abstract class VDiscardElement<T extends DiscardElement>
     extends VSvgElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VDiscardElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -155,11 +176,13 @@ class VEllipseElement extends VGeometryElement<EllipseElement> {
   EllipseElement elementFactory() => new EllipseElement();
 
   @override
+  @protected
   void applyAttributesToElement(EllipseElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VEllipseElement prev, EllipseElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -170,11 +193,13 @@ class VFEBlendElement extends VSvgElement<FEBlendElement> {
   FEBlendElement elementFactory() => new FEBlendElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEBlendElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFEBlendElement prev, FEBlendElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -185,11 +210,13 @@ class VFEColorMatrixElement extends VSvgElement<FEColorMatrixElement> {
   FEColorMatrixElement elementFactory() => new FEColorMatrixElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEColorMatrixElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFEColorMatrixElement prev, FEColorMatrixElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -203,11 +230,13 @@ class VFEComponentTransferElement
       new FEComponentTransferElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEComponentTransferElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFEComponentTransferElement prev, FEComponentTransferElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -217,11 +246,13 @@ class VFEComponentTransferElement
 abstract class VFECompositeElement<T extends FECompositeElement>
     extends VSvgElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VFECompositeElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -232,11 +263,13 @@ class VFEConvolveMatrixElement extends VSvgElement<FEConvolveMatrixElement> {
   FEConvolveMatrixElement elementFactory() => new FEConvolveMatrixElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEConvolveMatrixElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFEConvolveMatrixElement prev, FEConvolveMatrixElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -248,11 +281,13 @@ class VFEDiffuseLightingElement extends VSvgElement<FEDiffuseLightingElement> {
   FEDiffuseLightingElement elementFactory() => new FEDiffuseLightingElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEDiffuseLightingElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFEDiffuseLightingElement prev, FEDiffuseLightingElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -264,11 +299,13 @@ class VFEDisplacementMapElement extends VSvgElement<FEDisplacementMapElement> {
   FEDisplacementMapElement elementFactory() => new FEDisplacementMapElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEDisplacementMapElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFEDisplacementMapElement prev, FEDisplacementMapElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -280,11 +317,13 @@ class VFEDistantLightElement extends VSvgElement<FEDistantLightElement> {
   FEDistantLightElement elementFactory() => new FEDistantLightElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEDistantLightElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFEDistantLightElement prev, FEDistantLightElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -296,11 +335,13 @@ class VFEFloodElement extends VSvgElement<FEFloodElement> {
   FEFloodElement elementFactory() => new FEFloodElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEFloodElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFEFloodElement prev, FEFloodElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -311,11 +352,13 @@ class VFEFuncAElement extends VSvgElement<FEFuncAElement> {
   FEFuncAElement elementFactory() => new FEFuncAElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEFuncAElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFEFuncAElement prev, FEFuncAElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -326,11 +369,13 @@ class VFEFuncBElement extends VSvgElement<FEFuncBElement> {
   FEFuncBElement elementFactory() => new FEFuncBElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEFuncBElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFEFuncBElement prev, FEFuncBElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -341,11 +386,13 @@ class VFEFuncGElement extends VSvgElement<FEFuncGElement> {
   FEFuncGElement elementFactory() => new FEFuncGElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEFuncGElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFEFuncGElement prev, FEFuncGElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -356,11 +403,13 @@ class VFEFuncRElement extends VSvgElement<FEFuncRElement> {
   FEFuncRElement elementFactory() => new FEFuncRElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEFuncRElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFEFuncRElement prev, FEFuncRElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -371,11 +420,13 @@ class VFEGaussianBlurElement extends VSvgElement<FEGaussianBlurElement> {
   FEGaussianBlurElement elementFactory() => new FEGaussianBlurElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEGaussianBlurElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFEGaussianBlurElement prev, FEGaussianBlurElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -387,11 +438,13 @@ class VFEImageElement extends VSvgElement<FEImageElement> {
   FEImageElement elementFactory() => new FEImageElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEImageElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFEImageElement prev, FEImageElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -402,11 +455,13 @@ class VFEMergeElement extends VSvgElement<FEMergeElement> {
   FEMergeElement elementFactory() => new FEMergeElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEMergeElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFEMergeElement prev, FEMergeElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -417,11 +472,13 @@ class VFEMergeNodeElement extends VSvgElement<FEMergeNodeElement> {
   FEMergeNodeElement elementFactory() => new FEMergeNodeElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEMergeNodeElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFEMergeNodeElement prev, FEMergeNodeElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -431,11 +488,13 @@ class VFEMergeNodeElement extends VSvgElement<FEMergeNodeElement> {
 abstract class VFEMorphologyElement<T extends FEMorphologyElement>
     extends VSvgElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VFEMorphologyElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -446,11 +505,13 @@ class VFEOffsetElement extends VSvgElement<FEOffsetElement> {
   FEOffsetElement elementFactory() => new FEOffsetElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEOffsetElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFEOffsetElement prev, FEOffsetElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -461,11 +522,13 @@ class VFEPointLightElement extends VSvgElement<FEPointLightElement> {
   FEPointLightElement elementFactory() => new FEPointLightElement();
 
   @override
+  @protected
   void applyAttributesToElement(FEPointLightElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFEPointLightElement prev, FEPointLightElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -478,11 +541,13 @@ class VFESpecularLightingElement
   FESpecularLightingElement elementFactory() => new FESpecularLightingElement();
 
   @override
+  @protected
   void applyAttributesToElement(FESpecularLightingElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFESpecularLightingElement prev, FESpecularLightingElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -494,11 +559,13 @@ class VFESpotLightElement extends VSvgElement<FESpotLightElement> {
   FESpotLightElement elementFactory() => new FESpotLightElement();
 
   @override
+  @protected
   void applyAttributesToElement(FESpotLightElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFESpotLightElement prev, FESpotLightElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -510,11 +577,13 @@ class VFETileElement extends VSvgElement<FETileElement> {
   FETileElement elementFactory() => new FETileElement();
 
   @override
+  @protected
   void applyAttributesToElement(FETileElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFETileElement prev, FETileElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -525,11 +594,13 @@ class VFETurbulenceElement extends VSvgElement<FETurbulenceElement> {
   FETurbulenceElement elementFactory() => new FETurbulenceElement();
 
   @override
+  @protected
   void applyAttributesToElement(FETurbulenceElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VFETurbulenceElement prev, FETurbulenceElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -541,11 +612,13 @@ class VFilterElement extends VSvgElement<FilterElement> {
   FilterElement elementFactory() => new FilterElement();
 
   @override
+  @protected
   void applyAttributesToElement(FilterElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VFilterElement prev, FilterElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -556,11 +629,13 @@ class VForeignObjectElement extends VGraphicsElement<ForeignObjectElement> {
   ForeignObjectElement elementFactory() => new ForeignObjectElement();
 
   @override
+  @protected
   void applyAttributesToElement(ForeignObjectElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VForeignObjectElement prev, ForeignObjectElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -572,11 +647,13 @@ class VGElement extends VGraphicsElement<GElement> {
   GElement elementFactory() => new GElement();
 
   @override
+  @protected
   void applyAttributesToElement(GElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VGElement prev, GElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -585,11 +662,13 @@ class VGElement extends VGraphicsElement<GElement> {
 abstract class VGeometryElement<T extends GeometryElement>
     extends VGraphicsElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VGeometryElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -598,11 +677,13 @@ abstract class VGeometryElement<T extends GeometryElement>
 abstract class VGraphicsElement<T extends GraphicsElement>
     extends VSvgElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VGraphicsElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -613,11 +694,13 @@ class VImageElement extends VGraphicsElement<ImageElement> {
   ImageElement elementFactory() => new ImageElement();
 
   @override
+  @protected
   void applyAttributesToElement(ImageElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VImageElement prev, ImageElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -628,11 +711,13 @@ class VLineElement extends VGeometryElement<LineElement> {
   LineElement elementFactory() => new LineElement();
 
   @override
+  @protected
   void applyAttributesToElement(LineElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VLineElement prev, LineElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -643,11 +728,13 @@ class VLinearGradientElement extends VSvgElement<LinearGradientElement> {
   LinearGradientElement elementFactory() => new LinearGradientElement();
 
   @override
+  @protected
   void applyAttributesToElement(LinearGradientElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VLinearGradientElement prev, LinearGradientElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -659,11 +746,13 @@ class VMarkerElement extends VSvgElement<MarkerElement> {
   MarkerElement elementFactory() => new MarkerElement();
 
   @override
+  @protected
   void applyAttributesToElement(MarkerElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VMarkerElement prev, MarkerElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -674,11 +763,13 @@ class VMaskElement extends VSvgElement<MaskElement> {
   MaskElement elementFactory() => new MaskElement();
 
   @override
+  @protected
   void applyAttributesToElement(MaskElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VMaskElement prev, MaskElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -687,11 +778,13 @@ class VMaskElement extends VSvgElement<MaskElement> {
 abstract class VMetadataElement<T extends MetadataElement>
     extends VSvgElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VMetadataElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -702,11 +795,13 @@ class VPathElement extends VGeometryElement<PathElement> {
   PathElement elementFactory() => new PathElement();
 
   @override
+  @protected
   void applyAttributesToElement(PathElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VPathElement prev, PathElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -717,11 +812,13 @@ class VPatternElement extends VSvgElement<PatternElement> {
   PatternElement elementFactory() => new PatternElement();
 
   @override
+  @protected
   void applyAttributesToElement(PatternElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VPatternElement prev, PatternElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -732,11 +829,13 @@ class VPolygonElement extends VGeometryElement<PolygonElement> {
   PolygonElement elementFactory() => new PolygonElement();
 
   @override
+  @protected
   void applyAttributesToElement(PolygonElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VPolygonElement prev, PolygonElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -747,11 +846,13 @@ class VPolylineElement extends VGeometryElement<PolylineElement> {
   PolylineElement elementFactory() => new PolylineElement();
 
   @override
+  @protected
   void applyAttributesToElement(PolylineElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VPolylineElement prev, PolylineElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -762,11 +863,13 @@ class VRadialGradientElement extends VSvgElement<RadialGradientElement> {
   RadialGradientElement elementFactory() => new RadialGradientElement();
 
   @override
+  @protected
   void applyAttributesToElement(RadialGradientElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       VRadialGradientElement prev, RadialGradientElement ele) {
     super.updateElementAttributes(prev, ele);
@@ -778,11 +881,13 @@ class VRectElement extends VGeometryElement<RectElement> {
   RectElement elementFactory() => new RectElement();
 
   @override
+  @protected
   void applyAttributesToElement(RectElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VRectElement prev, RectElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -801,6 +906,7 @@ class VScriptElement extends VSvgElement<ScriptElement> {
   }
 
   @override
+  @protected
   void applyAttributesToElement(ScriptElement ele) {
     super.applyAttributesToElement(ele);
 
@@ -808,6 +914,7 @@ class VScriptElement extends VSvgElement<ScriptElement> {
   }
 
   @override
+  @protected
   void updateElementAttributes(VScriptElement prev, ScriptElement ele) {
     super.updateElementAttributes(prev, ele);
 
@@ -820,11 +927,13 @@ class VSetElement extends VAnimationElement<SetElement> {
   SetElement elementFactory() => new SetElement();
 
   @override
+  @protected
   void applyAttributesToElement(SetElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VSetElement prev, SetElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -835,11 +944,13 @@ class VStopElement extends VSvgElement<StopElement> {
   StopElement elementFactory() => new StopElement();
 
   @override
+  @protected
   void applyAttributesToElement(StopElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VStopElement prev, StopElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -874,6 +985,7 @@ class VStyleElement extends VSvgElement<StyleElement> {
   }
 
   @override
+  @protected
   void applyAttributesToElement(StyleElement ele) {
     super.applyAttributesToElement(ele);
 
@@ -883,6 +995,7 @@ class VStyleElement extends VSvgElement<StyleElement> {
   }
 
   @override
+  @protected
   void updateElementAttributes(VStyleElement prev, StyleElement ele) {
     super.updateElementAttributes(prev, ele);
 
@@ -902,6 +1015,7 @@ abstract class VSvgElement<T extends SvgElement> extends VElement<T> {
   }
 
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
 
@@ -909,6 +1023,7 @@ abstract class VSvgElement<T extends SvgElement> extends VElement<T> {
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VSvgElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
 
@@ -937,6 +1052,7 @@ class VSvgSvgElement extends VGraphicsElement<SvgSvgElement> {
   }
 
   @override
+  @protected
   void applyAttributesToElement(SvgSvgElement ele) {
     super.applyAttributesToElement(ele);
 
@@ -945,6 +1061,7 @@ class VSvgSvgElement extends VGraphicsElement<SvgSvgElement> {
   }
 
   @override
+  @protected
   void updateElementAttributes(VSvgSvgElement prev, SvgSvgElement ele) {
     super.updateElementAttributes(prev, ele);
 
@@ -958,11 +1075,13 @@ class VSwitchElement extends VGraphicsElement<SwitchElement> {
   SwitchElement elementFactory() => new SwitchElement();
 
   @override
+  @protected
   void applyAttributesToElement(SwitchElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VSwitchElement prev, SwitchElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -973,11 +1092,13 @@ class VSymbolElement extends VSvgElement<SymbolElement> {
   SymbolElement elementFactory() => new SymbolElement();
 
   @override
+  @protected
   void applyAttributesToElement(SymbolElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VSymbolElement prev, SymbolElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -988,11 +1109,13 @@ class VTSpanElement extends VTextPositioningElement<TSpanElement> {
   TSpanElement elementFactory() => new TSpanElement();
 
   @override
+  @protected
   void applyAttributesToElement(TSpanElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VTSpanElement prev, TSpanElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -1001,11 +1124,13 @@ class VTSpanElement extends VTextPositioningElement<TSpanElement> {
 abstract class VTextContentElement<T extends TextContentElement>
     extends VGraphicsElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VTextContentElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -1016,11 +1141,13 @@ class VTextElement extends VTextPositioningElement<TextElement> {
   TextElement elementFactory() => new TextElement();
 
   @override
+  @protected
   void applyAttributesToElement(TextElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VTextElement prev, TextElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -1029,11 +1156,13 @@ class VTextElement extends VTextPositioningElement<TextElement> {
 abstract class VTextPathElement<T extends TextPathElement>
     extends VTextContentElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(covariant VTextPathElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -1042,11 +1171,13 @@ abstract class VTextPathElement<T extends TextPathElement>
 abstract class VTextPositioningElement<T extends TextPositioningElement>
     extends VTextContentElement<T> {
   @override
+  @protected
   void applyAttributesToElement(T ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(
       covariant VTextPositioningElement<T> prev, T ele) {
     super.updateElementAttributes(prev, ele);
@@ -1058,11 +1189,13 @@ class VTitleElement extends VSvgElement<TitleElement> {
   TitleElement elementFactory() => new TitleElement();
 
   @override
+  @protected
   void applyAttributesToElement(TitleElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VTitleElement prev, TitleElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -1073,11 +1206,13 @@ class VUseElement extends VGraphicsElement<UseElement> {
   UseElement elementFactory() => new UseElement();
 
   @override
+  @protected
   void applyAttributesToElement(UseElement ele) {
     super.applyAttributesToElement(ele);
   }
 
   @override
+  @protected
   void updateElementAttributes(VUseElement prev, UseElement ele) {
     super.updateElementAttributes(prev, ele);
   }
@@ -1096,6 +1231,7 @@ class VViewElement extends VSvgElement<ViewElement> {
   }
 
   @override
+  @protected
   void applyAttributesToElement(ViewElement ele) {
     super.applyAttributesToElement(ele);
 
@@ -1103,6 +1239,7 @@ class VViewElement extends VSvgElement<ViewElement> {
   }
 
   @override
+  @protected
   void updateElementAttributes(VViewElement prev, ViewElement ele) {
     super.updateElementAttributes(prev, ele);
