+++ lib/wui_builder.dart
@@ -1,13 +1,7 @@
 library wui_builder;
 
-import 'dart:html';
-import 'package:meta/meta.dart';
-import 'vhtml.dart' show VElement;
-
-part 'src/wui_builder/component.dart';
-part 'src/wui_builder/create_nodes.dart';
-part 'src/wui_builder/cursors.dart';
-part 'src/wui_builder/update_tracker.dart';
-part 'src/wui_builder/update_queue.dart';
-part 'src/wui_builder/update_processor.dart';
-part 'src/wui_builder/vnode.dart';
+export 'src/wui_builder/create_nodes.dart' show render;
+export 'src/wui_builder/component.dart'
+    show Component, PropComponent, StateSetter;
+export 'src/wui_builder/velement.dart' show VElement;
+export 'src/wui_builder/vnode.dart' show VNode;
